First working version of the DDSA.
Forked from Swarmathon-ROS. Added a DDSA Controller class to contain the DDSA logic. Changed the Swarmathon-ROS from a binary left and right rotational controller to a proportional control.
The logic in the DDSA controller is the same as in the ARGoS implementation.
Only one rover is spawned in the prelim case for testing, instead of three.